+++ substreams-entity-change/src/lib.rs
@@ -1,3 +1,5 @@
+extern crate core;
+
 pub mod change;
 pub mod helpers;
 pub mod pb;